Make ImagePlaceholder a reusable component
